reorganize policy related code
